requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.19.0
jira>=3.4.0
matplotlib>=3.5.0
seaborn>=0.12.0